+++ dev/Network.js
@@ -122,13 +122,13 @@ class Network {
                     this.initLayers()
                     break
 
-                case layers.every(item => item instanceof Layer):
+                case layers.every(item => item instanceof Layer || item instanceof FCLayer):
                     this.state = "constructed"
                     this.layers = layers
                     this.initLayers()
                     break
 
-                case layers.every(item => item === Layer):
+                case layers.every(item => item === Layer || item === FCLayer):
                     this.state = "defined"
                     this.definedLayers = layers
                     break
@@ -165,10 +165,10 @@ class Network {
                 break
 
             case "not-defined":
-                this.layers[0] = new Layer(input)
-                this.layers[1] = new Layer(Math.ceil(input/expected > 5 ? expected + (Math.abs(input-expected))/4
+                this.layers[0] = new FCLayer(input)
+                this.layers[1] = new FCLayer(Math.ceil(input/expected > 5 ? expected + (Math.abs(input-expected))/4
                                                                         : input + expected))
-                this.layers[2] = new Layer(Math.ceil(expected))
+                this.layers[2] = new FCLayer(Math.ceil(expected))
                 break
         }
 
@@ -429,7 +429,7 @@ class Network {
             throw new Error("No JSON data given to import.")
         }
 
-        this.layers = data.layers.map(layer => new Layer(layer.neurons.length, layer.neurons))
+        this.layers = data.layers.map(layer => new FCLayer(layer.neurons.length, layer.neurons))
         this.state = "constructed"
         this.initLayers()
     }